Refactor to reuse common logic for quickfixes
