+++ src/IssueViz/Editor/QuickActions/QuickFixes/QuickFixActionsSource.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * SonarLint for Visual Studio
  * Copyright (C) SonarSource Sàrl
  * mailto:info AT sonarsource DOT com
@@ -20,132 +20,41 @@
 
 using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Tagging;
-using Microsoft.VisualStudio.Threading;
+using Microsoft.VisualStudio.Text.Editor;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Telemetry;
-using SonarLint.VisualStudio.IssueVisualization.Editor.LocationTagging;
 using SonarLint.VisualStudio.IssueVisualization.Models;
 
 namespace SonarLint.VisualStudio.IssueVisualization.Editor.QuickActions.QuickFixes;
 
-internal sealed class QuickFixActionsSource : ISuggestedActionsSource
+internal sealed class QuickFixActionsSource(
+    ILightBulbBroker lightBulbBroker,
+    IBufferTagAggregatorFactoryService bufferTagAggregatorFactoryService,
+    ITextView textView,
+    ITextBuffer textBuffer,
+    IQuickFixesTelemetryManager quickFixesTelemetryManager,
+    IMessageBox messageBox,
+    ILogger logger,
+    IThreadHandling threadHandling)
+    : IssueActionsSourceBase(lightBulbBroker, bufferTagAggregatorFactoryService, textView, textBuffer, logger, threadHandling)
 {
-    private readonly ILightBulbBroker lightBulbBroker;
-    private readonly ITextBuffer textBuffer;
-    private readonly ITextView textView;
-    private readonly ILogger logger;
-    private readonly IThreadHandling threadHandling;
-    private readonly ITagAggregator<IIssueLocationTag> issueLocationsTagAggregator;
-    private readonly IMessageBox messageBox;
-    private readonly IQuickFixesTelemetryManager quickFixesTelemetryManager;
+    private readonly ITextBuffer textBuffer = textBuffer;
+    private readonly IThreadHandling threadHandling = threadHandling;
 
-    public QuickFixActionsSource(ILightBulbBroker lightBulbBroker,
-        IBufferTagAggregatorFactoryService bufferTagAggregatorFactoryService,
-        ITextView textView,
-        ITextBuffer textBuffer,
-        IQuickFixesTelemetryManager quickFixesTelemetryManager,
-        IMessageBox messageBox,
-        ILogger logger,
-        IThreadHandling threadHandling)
+    protected override SuggestedActionSetPriority Priority => SuggestedActionSetPriority.Medium;
+
+    protected override bool TryGetMatchingIssues(IEnumerable<IAnalysisIssueVisualization> issueVisualizations, out IEnumerable<IAnalysisIssueVisualization> matchingIssues)
     {
-        this.lightBulbBroker = lightBulbBroker;
-        this.textBuffer = textBuffer;
-        this.textView = textView;
-        this.quickFixesTelemetryManager = quickFixesTelemetryManager;
-        this.messageBox = messageBox;
-        this.logger = logger;
-        this.threadHandling = threadHandling;
+        matchingIssues = issueVisualizations
+            .Where(x => x.QuickFixes.Any(fix => fix.CanBeApplied(textBuffer.CurrentSnapshot)));
 
-        issueLocationsTagAggregator = bufferTagAggregatorFactoryService.CreateTagAggregator<IIssueLocationTag>(textBuffer);
-        issueLocationsTagAggregator.TagsChanged += TagAggregator_TagsChanged;
+        return matchingIssues.Any();
     }
 
-    public event EventHandler<EventArgs> SuggestedActionsChanged;
-
-    public IEnumerable<SuggestedActionSet> GetSuggestedActions(
-        ISuggestedActionCategorySet requestedActionCategories,
-        SnapshotSpan range,
-        CancellationToken cancellationToken)
-    {
-        var allActions = new List<ISuggestedAction>();
-
-        try
-        {
-            if (IsOnIssueWithApplicableQuickFixes(range, out var issuesWithFixes))
-            {
-                foreach (var issueViz in issuesWithFixes)
-                {
-                    var applicableFixes = issueViz.QuickFixes.Where(x => x.CanBeApplied(textBuffer.CurrentSnapshot));
-
-                    allActions.AddRange(applicableFixes.Select(fix => new QuickFixSuggestedAction(fix, textBuffer, issueViz, quickFixesTelemetryManager, messageBox, logger, threadHandling)));
-                }
-            }
-        }
-        catch (Exception ex) when (!ErrorHandler.IsCriticalException(ex))
-        {
-            logger.WriteLine(string.Format(Resources.ERR_QuickFixes_Exception, ex));
-        }
-
-        return allActions.Any()
-            ? new[] { new SuggestedActionSet(allActions, priority: SuggestedActionSetPriority.Medium) }
-            : Enumerable.Empty<SuggestedActionSet>();
-    }
-
-    public async Task<bool> HasSuggestedActionsAsync(ISuggestedActionCategorySet requestedActionCategories, SnapshotSpan range, CancellationToken cancellationToken)
-    {
-        var hasActions = false;
-        try
-        {
-            await threadHandling.RunOnUIThreadAsync(() => hasActions = IsOnIssueWithApplicableQuickFixes(range, out _));
-        }
-        catch(Exception ex) when (!ErrorHandler.IsCriticalException(ex))
-        {
-            logger.WriteLine(string.Format(Resources.ERR_QuickFixes_Exception, ex));
-        }
-        return hasActions;
-    }
-
-    private bool IsOnIssueWithApplicableQuickFixes(SnapshotSpan range, out IEnumerable<IAnalysisIssueVisualization> issuesWithFixes)
-    {
-        var tagSpans = issueLocationsTagAggregator.GetTags(range);
-
-        issuesWithFixes = tagSpans
-            .Select(x => x.Tag.Location)
-            .OfType<IAnalysisIssueVisualization>()
-            .Where(x =>
-                x.QuickFixes.Any(fix => fix.CanBeApplied(textBuffer.CurrentSnapshot)));
-
-        return issuesWithFixes.Any();
-    }
-
-    public bool TryGetTelemetryId(out Guid telemetryId)
-    {
-        telemetryId = Guid.Empty;
-        return false;
-    }
-
-    public void Dispose()
-    {
-        issueLocationsTagAggregator.TagsChanged -= TagAggregator_TagsChanged;
-        issueLocationsTagAggregator.Dispose();
-    }
-
-    private void TagAggregator_TagsChanged(object sender, TagsChangedEventArgs e)
-        => HandleTagsChangedAsync().Forget();
-
-    internal /* for testing */ async Task HandleTagsChangedAsync()
-    {
-        try
-        {
-            await threadHandling.RunOnUIThreadAsync(() => lightBulbBroker.DismissSession(textView));
-
-            SuggestedActionsChanged?.Invoke(this, EventArgs.Empty);
-        }
-        catch (Exception ex) when (!ErrorHandler.IsCriticalException(ex))
-        {
-            logger.LogVerbose($"[QuickFixActionsSource] Exception handling TagsChanged event: {ex}");
-        }
-    }
+    protected override IEnumerable<ISuggestedAction> CreateActions(IEnumerable<IAnalysisIssueVisualization> matchingIssues) =>
+        matchingIssues.SelectMany(issueViz =>
+            issueViz.QuickFixes
+                .Where(x => x.CanBeApplied(textBuffer.CurrentSnapshot))
+                .Select(fix => new QuickFixSuggestedAction(fix, textBuffer, issueViz, quickFixesTelemetryManager, messageBox, logger, threadHandling)));
 }
